requirements.txt
Flask
Flask-Session
Flask-Login
Flask-SQLAlchemy
Werkzeug
flask-restful
flask-jwt-extended
flask-cors
celery==5.0.5
redis
flower

mmcv-full
mmdet
mmcls
onnx
onnx-simplifier